fix:修改销售订单延保

+++ OpenAuth.Core/OpenAuth.App/Material/Request/AddOrUpdateQuotationReq.cs
@@ -261,7 +261,7 @@ namespace OpenAuth.App.Material.Request
         /// <summary>
         ///延保类型 1更新 2销售
         /// </summary>
-        public decimal? WarrantyType { get; set; }
+        public int? WarrantyType { get; set; }
         /// <summary>
         /// 报价单设备列表
         /// </summary>
